+++ packages/viewer/web/src/components/Anchor.js
@@ -1,10 +1,13 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-const Anchor = React.memo(({ anchorKey }) => <div key={anchorKey} data-hook={anchorKey} />);
+const Anchor = React.memo(({ anchorKey, type }) => (
+  <div key={anchorKey} type={type} data-hook={anchorKey} />
+));
 
 Anchor.propTypes = {
   anchorKey: PropTypes.string,
+  type: PropTypes.string,
 };
 
 export default Anchor;